# crates/api-gateway/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Denial Navigator API",
    "version": "1.0.0",
    "description": "Self-hosted denial-management API (Rust / axum rewrite). Every browser request goes through this gateway; it enforces authentication, role-based authorisation, per-request account currency and HIPAA access auditing.\n\nAuth: `POST /api/v1/auth/login` returns a bearer token. When TOTP is required it returns an mfa-scope token usable only on the `/auth/totp/*`, `/auth/login/totp` and `/auth/me` paths until `POST /auth/login/totp` upgrades it. Sibling services present `X-Service-Key` instead."
  },
  "servers": [
    {
      "url": "/",
      "description": "This gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "serviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Key"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "summary": "Audit log entries (manager+)",
        "tags": [
          "audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Actor name, incl. 'service:*' / 'anonymous'."
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Max rows to return."
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/retention/audit": {
      "get": {
        "summary": "Audit-log size, age and what a prune would remove (admin)",
        "tags": [
          "retention"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/retention/audit/prune": {
      "post": {
        "summary": "Delete audit entries past the retention window (admin)",
        "tags": [
          "retention"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires confirm=true; refuses a window under a year; logs an 'audit_pruned' entry after deleting.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "older_than_days": {
                    "type": "integer"
                  },
                  "confirm": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "confirm"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}